+++ replacer-backend/src/main/java/es/bvalero/replacer/common/domain/ReplacementKind.java
@@ -14,7 +14,6 @@ public enum ReplacementKind {
     SIMPLE((byte) 2),
     COMPOSED((byte) 3),
     CUSTOM((byte) 1),
-    DATE((byte) 4),
     STYLE((byte) 5),
     EMPTY((byte) 99); // To be used but not meant to be serialized