+++ replacer-backend/src/test/java/es/bvalero/replacer/finder/replacement/finders/DateFinderTest.java
@@ -20,14 +20,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest(classes = { DateFinder.class, XmlConfiguration.class })
 class DateFinderTest {
 
-    // The CSV source doesn't allow the value to be non-constant
-    // even if we take the subtype from the replacement type constants
-    private static final String SUBTYPE_DOT_YEAR = "Año con punto";
-    private static final String SUBTYPE_INCOMPLETE = "Fecha incompleta";
-    private static final String SUBTYPE_LEADING_ZERO = "Día con cero";
-    private static final String SUBTYPE_UPPERCASE = "Mes en mayúscula";
-    private static final String SUBTYPE_UNORDERED = "Fecha desordenada";
-
     @Autowired
     private DateFinder dateFinder;
 
@@ -35,43 +27,43 @@ class DateFinderTest {
     @CsvSource(
         delimiter = '|',
         value = {
-            "7 de Agosto de 2019|7 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "7 de  Agosto 2019|7 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 de Agosto de 2019|17 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "07 de Agosto de 2019|7 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 de Agosto del 2019|17 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 de Agosto de 2.019|17 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 Agosto de 2019|17 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 de Agosto 2019|17 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 Agosto 2019|17 de agosto de 2019|" + SUBTYPE_UPPERCASE,
-            "17 de Setiembre de 2019|17 de septiembre de 2019|" + SUBTYPE_UPPERCASE,
-            "07 de agosto de 2019|7 de agosto de 2019|" + SUBTYPE_LEADING_ZERO,
-            "07 de agosto del 2019|7 de agosto de 2019|" + SUBTYPE_LEADING_ZERO,
-            "07 agosto de 2019|7 de agosto de 2019|" + SUBTYPE_LEADING_ZERO,
-            "07 de agosto 2019|7 de agosto de 2019|" + SUBTYPE_LEADING_ZERO,
-            "07 agosto 2019|7 de agosto de 2019|" + SUBTYPE_LEADING_ZERO,
-            "07 de agosto de 2.019|7 de agosto de 2019|" + SUBTYPE_LEADING_ZERO,
-            "07 de setiembre de 2019|7 de septiembre de 2019|" + SUBTYPE_LEADING_ZERO,
-            "17 de agosto 2019|17 de agosto de 2019|" + SUBTYPE_INCOMPLETE,
-            "17 agosto de 2019|17 de agosto de 2019|" + SUBTYPE_INCOMPLETE,
-            "17 agosto del 2019|17 de agosto de 2019|" + SUBTYPE_INCOMPLETE,
-            "17 agosto 2019|17 de agosto de 2019|" + SUBTYPE_INCOMPLETE,
-            "17 agosto 2.019|17 de agosto de 2019|" + SUBTYPE_DOT_YEAR,
-            "17 setiembre 2019|17 de septiembre de 2019|" + SUBTYPE_INCOMPLETE,
-            "17 de agosto de 2.019|17 de agosto de 2019|" + SUBTYPE_DOT_YEAR,
-            "17 de agosto del 2.019|17 de agosto de 2019|" + SUBTYPE_DOT_YEAR,
-            "17 de setiembre de 2.019|17 de septiembre de 2019|" + SUBTYPE_DOT_YEAR,
-            "10 de septiembre, 2022|10 de septiembre de 2022|" + SUBTYPE_INCOMPLETE,
+            "7 de Agosto de 2019|7 de agosto de 2019",
+            "7 de  Agosto 2019|7 de agosto de 2019",
+            "17 de Agosto de 2019|17 de agosto de 2019",
+            "07 de Agosto de 2019|7 de agosto de 2019",
+            "17 de Agosto del 2019|17 de agosto de 2019",
+            "17 de Agosto de 2.019|17 de agosto de 2019",
+            "17 Agosto de 2019|17 de agosto de 2019",
+            "17 de Agosto 2019|17 de agosto de 2019",
+            "17 Agosto 2019|17 de agosto de 2019",
+            "17 de Setiembre de 2019|17 de septiembre de 2019",
+            "07 de agosto de 2019|7 de agosto de 2019",
+            "07 de agosto del 2019|7 de agosto de 2019",
+            "07 agosto de 2019|7 de agosto de 2019",
+            "07 de agosto 2019|7 de agosto de 2019",
+            "07 agosto 2019|7 de agosto de 2019",
+            "07 de agosto de 2.019|7 de agosto de 2019",
+            "07 de setiembre de 2019|7 de septiembre de 2019",
+            "17 de agosto 2019|17 de agosto de 2019",
+            "17 agosto de 2019|17 de agosto de 2019",
+            "17 agosto del 2019|17 de agosto de 2019",
+            "17 agosto 2019|17 de agosto de 2019",
+            "17 agosto 2.019|17 de agosto de 2019",
+            "17 setiembre 2019|17 de septiembre de 2019",
+            "17 de agosto de 2.019|17 de agosto de 2019",
+            "17 de agosto del 2.019|17 de agosto de 2019",
+            "17 de setiembre de 2.019|17 de septiembre de 2019",
+            "10 de septiembre, 2022|10 de septiembre de 2022",
         }
     )
-    void testLongDate(String date, String expected, String subtype) {
+    void testLongDate(String date, String expected) {
         List<Replacement> replacements = dateFinder.findList(date);
 
         assertEquals(1, replacements.size());
         assertEquals(date, replacements.get(0).getText());
         assertEquals(date, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(expected, replacements.get(0).getSuggestions().get(1).getText());
-        assertEquals(subtype, replacements.get(0).getType().getSubtype());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 
     @ParameterizedTest
@@ -113,25 +105,25 @@ class DateFinderTest {
     @ParameterizedTest
     @CsvSource(
         {
-            "Desde Agosto de 2019, Desde agosto de 2019, " + SUBTYPE_UPPERCASE,
-            "desde Agosto de 2019, desde agosto de 2019, " + SUBTYPE_UPPERCASE,
-            "desde Agosto del 2019, desde agosto de 2019, " + SUBTYPE_UPPERCASE,
-            "desde Agosto de 2.019, desde agosto de 2019, " + SUBTYPE_UPPERCASE,
-            "desde Agosto 2019, desde agosto de 2019, " + SUBTYPE_UPPERCASE,
-            "Desde agosto 2019, Desde agosto de 2019, " + SUBTYPE_INCOMPLETE,
-            "desde agosto 2.019, desde agosto de 2019, " + SUBTYPE_DOT_YEAR,
-            "Desde agosto de 2.019, Desde agosto de 2019, " + SUBTYPE_DOT_YEAR,
-            "En agosto del 2.019, En agosto de 2019, " + SUBTYPE_DOT_YEAR,
+            "Desde Agosto de 2019, Desde agosto de 2019",
+            "desde Agosto de 2019, desde agosto de 2019",
+            "desde Agosto del 2019, desde agosto de 2019",
+            "desde Agosto de 2.019, desde agosto de 2019",
+            "desde Agosto 2019, desde agosto de 2019",
+            "Desde agosto 2019, Desde agosto de 2019",
+            "desde agosto 2.019, desde agosto de 2019",
+            "Desde agosto de 2.019, Desde agosto de 2019",
+            "En agosto del 2.019, En agosto de 2019",
         }
     )
-    void testMonthYear(String date, String expected, String subtype) {
+    void testMonthYear(String date, String expected) {
         List<Replacement> replacements = dateFinder.findList(date);
 
         assertEquals(1, replacements.size());
         assertEquals(date, replacements.get(0).getText());
         assertEquals(date, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(expected, replacements.get(0).getSuggestions().get(1).getText());
-        assertEquals(subtype, replacements.get(0).getType().getSubtype());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 
     @ParameterizedTest
@@ -163,30 +155,30 @@ class DateFinderTest {
         assertEquals(date, replacements.get(0).getText());
         assertEquals(date, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(expected, replacements.get(0).getSuggestions().get(1).getText());
-        assertEquals(SUBTYPE_INCOMPLETE, replacements.get(0).getType().getSubtype());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 
     @ParameterizedTest
     @CsvSource(
         delimiter = '|',
         value = {
-            "Mayo 3, 1999|3 de mayo de 1999|" + SUBTYPE_UNORDERED,
-            "Mayo 3 1999|3 de mayo de 1999|" + SUBTYPE_UNORDERED,
-            "Octubre 14 de 2006|14 de octubre de 2006|" + SUBTYPE_UNORDERED,
-            "Octubre 14 del 2006|14 de octubre de 2006|" + SUBTYPE_UNORDERED,
-            "Mayo 03, 1999|3 de mayo de 1999|" + SUBTYPE_UNORDERED,
-            "Mayo 3, 1.999|3 de mayo de 1999|" + SUBTYPE_UNORDERED,
-            "mayo 3, 1999|3 de mayo de 1999|" + SUBTYPE_UNORDERED,
+            "Mayo 3, 1999|3 de mayo de 1999",
+            "Mayo 3 1999|3 de mayo de 1999",
+            "Octubre 14 de 2006|14 de octubre de 2006",
+            "Octubre 14 del 2006|14 de octubre de 2006",
+            "Mayo 03, 1999|3 de mayo de 1999",
+            "Mayo 3, 1.999|3 de mayo de 1999",
+            "mayo 3, 1999|3 de mayo de 1999",
         }
     )
-    void testUnorderedDateMonthDayYear(String date, String expected, String subtype) {
+    void testUnorderedDateMonthDayYear(String date, String expected) {
         List<Replacement> replacements = dateFinder.findList(date);
 
         assertEquals(1, replacements.size());
         assertEquals(date, replacements.get(0).getText());
         assertEquals(date, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(expected, replacements.get(0).getSuggestions().get(1).getText());
-        assertEquals(subtype, replacements.get(0).getType().getSubtype());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 
     @ParameterizedTest
@@ -204,29 +196,29 @@ class DateFinderTest {
         assertEquals(match, replacements.get(0).getText());
         assertEquals(match, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(fix, replacements.get(0).getSuggestions().get(1).getText());
-        assertEquals(ReplacementType.DATE_UNORDERED, replacements.get(0).getType());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 
     @ParameterizedTest
     @CsvSource(
         delimiter = '|',
         value = {
-            "2013, Octubre 30|30 de octubre de 2013|" + SUBTYPE_UNORDERED,
-            "2013 Octubre 30|30 de octubre de 2013|" + SUBTYPE_UNORDERED,
-            "2013, octubre 30|30 de octubre de 2013|" + SUBTYPE_UNORDERED,
-            "2013, Mayo 4|4 de mayo de 2013|" + SUBTYPE_UNORDERED,
-            "2013, Mayo 04|4 de mayo de 2013|" + SUBTYPE_UNORDERED,
-            "2.013, Mayo 4|4 de mayo de 2013|" + SUBTYPE_UNORDERED,
+            "2013, Octubre 30|30 de octubre de 2013",
+            "2013 Octubre 30|30 de octubre de 2013",
+            "2013, octubre 30|30 de octubre de 2013",
+            "2013, Mayo 4|4 de mayo de 2013",
+            "2013, Mayo 04|4 de mayo de 2013",
+            "2.013, Mayo 4|4 de mayo de 2013",
         }
     )
-    void testUnorderedDateYearMonthDay(String date, String expected, String subtype) {
+    void testUnorderedDateYearMonthDay(String date, String expected) {
         List<Replacement> replacements = dateFinder.findList(date);
 
         assertEquals(1, replacements.size());
         assertEquals(date, replacements.get(0).getText());
         assertEquals(date, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(expected, replacements.get(0).getSuggestions().get(1).getText());
-        assertEquals(subtype, replacements.get(0).getType().getSubtype());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 
     @ParameterizedTest
@@ -249,12 +241,12 @@ class DateFinderTest {
     @CsvSource(
         delimiter = '|',
         value = {
-            "en 22 de Agosto de 2022|el 22 de agosto de 2022|en 22 de agosto de 2022|" + SUBTYPE_UPPERCASE,
-            "En Agosto 22, 2022|El 22 de agosto de 2022|En 22 de agosto de 2022|" + SUBTYPE_UNORDERED,
-            "A 2022, Agosto 22|Al 22 de agosto de 2022|A 22 de agosto de 2022|" + SUBTYPE_UNORDERED,
+            "en 22 de Agosto de 2022|el 22 de agosto de 2022|en 22 de agosto de 2022",
+            "En Agosto 22, 2022|El 22 de agosto de 2022|En 22 de agosto de 2022",
+            "A 2022, Agosto 22|Al 22 de agosto de 2022|A 22 de agosto de 2022",
         }
     )
-    void testNotValidDateWithWrongArticle(String date, String expected1, String expected2, String subtype) {
+    void testNotValidDateWithWrongArticle(String date, String expected1, String expected2) {
         List<Replacement> replacements = dateFinder.findList(date);
 
         assertEquals(1, replacements.size());
@@ -263,6 +255,6 @@ class DateFinderTest {
         assertEquals(date, replacements.get(0).getSuggestions().get(0).getText());
         assertEquals(expected1, replacements.get(0).getSuggestions().get(1).getText());
         assertEquals(expected2, replacements.get(0).getSuggestions().get(2).getText());
-        assertEquals(subtype, replacements.get(0).getType().getSubtype());
+        assertEquals(ReplacementType.DATE, replacements.get(0).getType());
     }
 }